+++ src/react/src/components/ProductionLocation/ProductionLocationDetails/ProductionLocationDetails.jsx
@@ -1,11 +1,13 @@
 import React from 'react';
 import { connect } from 'react-redux';
+import { withRouter } from 'react-router';
 import { withStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
 import ArrowBack from '@material-ui/icons/ArrowBackIos';
 
 import { resetSingleFacility } from '../../../actions/facilities';
 import { facilitiesRoute } from '../../../util/constants';
+import { shouldUseProductionLocationPage } from '../../../util/util';
 
 import ProductionLocationDetailsContent from '../ProductionLocationDetailsContent/ProductionLocationDetailsContent';
 
@@ -15,6 +17,10 @@ function ProductionLocationDetails({
     classes,
     clearFacility,
     history: { push },
+    data,
+    embed,
+    useProductionLocationPage,
+    location,
 }) {
     return (
         <div className={classes.container}>
@@ -32,13 +38,32 @@ function ProductionLocationDetails({
                 </Button>
             </div>
             <p>Production Location Details</p>
-            <ProductionLocationDetailsContent />
+            <ProductionLocationDetailsContent
+                data={data}
+                embed={embed}
+                clearFacility={clearFacility}
+                useProductionLocationPage={useProductionLocationPage}
+                location={location}
+            />
         </div>
     );
 }
 
-function mapStateToProps({ filters, embeddedMap: { embed } }) {
-    return { filters, embedded: !!embed };
+function mapStateToProps({
+    filters,
+    embeddedMap: { embed },
+    facilities: { singleFacility: { data } = {} } = {},
+    featureFlags,
+}) {
+    return {
+        filters,
+        embedded: !!embed,
+        data,
+        embed,
+        useProductionLocationPage: shouldUseProductionLocationPage(
+            featureFlags,
+        ),
+    };
 }
 
 function mapDispatchToProps(dispatch) {
@@ -47,7 +72,9 @@ function mapDispatchToProps(dispatch) {
     };
 }
 
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps,
-)(withStyles(productionLocationDetailsStyles)(ProductionLocationDetails));
+export default withRouter(
+    connect(
+        mapStateToProps,
+        mapDispatchToProps,
+    )(withStyles(productionLocationDetailsStyles)(ProductionLocationDetails)),
+);